📝 Update CondenseSkill to use separator variable
The CondenseSkill was modified to use a variable for the separator
string instead of a constant. This allows for more flexibility and
customization when condensing different types of inputs. The separator
variable is set and updated in the CondenseSkill.cs file, and passed
to the Condenser semantic function. The skprompt.txt file was also
updated to reflect the changes in the input and output format.